moved rusky_pvp.lua to gen_misc
